# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Punjabi Kids Gamification API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /spinner/configs:
    get:
      operationId: listSpinnerConfigs
      tags: [spinner]
      summary: List all spinner configurations
      responses:
        "200":
          description: List of spinner configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpinnerConfig"
    post:
      operationId: createSpinnerConfig
      tags: [spinner]
      summary: Create a spinner configuration (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpinnerConfigInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpinnerConfig"
  /spinner/configs/{id}:
    get:
      operationId: getSpinnerConfig
      tags: [spinner]
      summary: Get a spinner configuration by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Spinner config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpinnerConfig"
    patch:
      operationId: updateSpinnerConfig
      tags: [spinner]
      summary: Update a spinner configuration (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpinnerConfigUpdate"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpinnerConfig"
    delete:
      operationId: deleteSpinnerConfig
      tags: [spinner]
      summary: Delete a spinner configuration (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /spinner/spin:
    post:
      operationId: recordSpin
      tags: [spinner]
      summary: Record a spin result and award points
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpinResult"
      responses:
        "200":
          description: Spin recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpinRecord"

  # ─── TONGUE TWISTER ──────────────────────────────────────────
  /tongue-twisters:
    get:
      operationId: listTongueTwisters
      tags: [tongue-twister]
      summary: List all tongue twisters
      responses:
        "200":
          description: List of tongue twisters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TongueTwister"
    post:
      operationId: createTongueTwister
      tags: [tongue-twister]
      summary: Create a tongue twister (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TongueTwisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TongueTwister"
  /tongue-twisters/{id}:
    get:
      operationId: getTongueTwister
      tags: [tongue-twister]
      summary: Get a tongue twister by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tongue twister
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TongueTwister"
    patch:
      operationId: updateTongueTwister
      tags: [tongue-twister]
      summary: Update a tongue twister (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TongueTwisterUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TongueTwister"
    delete:
      operationId: deleteTongueTwister
      tags: [tongue-twister]
      summary: Delete a tongue twister (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tongue-twisters/submissions:
    get:
      operationId: listTongueTwisterSubmissions
      tags: [tongue-twister]
      summary: List voice recording submissions (admin can see all, user sees own)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TongueTwisterSubmission"
  /tongue-twisters/{id}/submit:
    post:
      operationId: submitTongueTwisterRecording
      tags: [tongue-twister]
      summary: Submit a voice recording for a tongue twister
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TongueTwisterRecordingInput"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TongueTwisterSubmission"
  /tongue-twisters/submissions/{id}/review:
    patch:
      operationId: reviewTongueTwisterSubmission
      tags: [tongue-twister]
      summary: Admin reviews and scores a submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionReview"
      responses:
        "200":
          description: Reviewed submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TongueTwisterSubmission"

  # ─── KNOWLEDGE TEST ──────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
            enum: [user, admin, moderator]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [admin]
      summary: Update a user's role (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── UPLOADS ─────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, username, role, totalPoints, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin, moderator]
        totalPoints:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    RoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [user, admin, moderator]

    # ─── SPINNER ────────────────────────────────────────────────
    SpinnerItem:
      type: object
      required: [label, type]
      properties:
        label:
          type: string
        type:
          type: string
          enum: [text, image, number]
        imageUrl:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        weight:
          type: ["integer", "null"]
    SpinnerConfig:
      type: object
      required: [id, name, items, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SpinnerItem"
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    SpinnerConfigInput:
      type: object
      required: [name, items]
      properties:
        name:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SpinnerItem"
        isActive:
          type: boolean
    SpinnerConfigUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SpinnerItem"
        isActive:
          type: boolean
    SpinResult:
      type: object
      required: [configId, resultLabel]
      properties:
        configId:
          type: integer
        resultLabel:
          type: string
    SpinRecord:
      type: object
      required: [id, configId, resultLabel, pointsEarned, createdAt]
      properties:
        id:
          type: integer
        configId:
          type: integer
        resultLabel:
          type: string
        pointsEarned:
          type: integer
        createdAt:
          type: string
          format: date-time

    # ─── TONGUE TWISTER ─────────────────────────────────────────
    TongueTwister:
      type: object
      required: [id, text, transliteration, difficulty, createdAt]
      properties:
        id:
          type: integer
        text:
          type: string
        transliteration:
          type: ["string", "null"]
        translation:
          type: ["string", "null"]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    TongueTwisterInput:
      type: object
      required: [text, difficulty]
      properties:
        text:
          type: string
        transliteration:
          type: string
        translation:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        isActive:
          type: boolean
    TongueTwisterUpdate:
      type: object
      properties:
        text:
          type: string
        transliteration:
          type: string
        translation:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        isActive:
          type: boolean
    TongueTwisterRecordingInput:
      type: object
      required: [audioUrl]
      properties:
        audioUrl:
          type: string
        durationSeconds:
          type: ["number", "null"]
    TongueTwisterSubmission:
      type: object
      required: [id, tongueTwisterId, userId, audioUrl, status, createdAt]
      properties:
        id:
          type: integer
        tongueTwisterId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        audioUrl:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        score:
          type: ["integer", "null"]
        feedback:
          type: ["string", "null"]
        pointsEarned:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        reviewedAt:
          type: ["string", "null"]
          format: date-time
    SubmissionReview:
      type: object
      required: [status, score]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        score:
          type: integer
          minimum: 0
          maximum: 100
        feedback:
          type: string

    # ─── KNOWLEDGE TEST ─────────────────────────────────────────
...