# lib/api-zod/src/generated/types/tongueTwister.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Punjabi Kids Gamification API
 * OpenAPI spec version: 0.1.0
 */
import type { TongueTwisterDifficulty } from './tongueTwisterDifficulty';

export interface TongueTwister {
  id: number;
  text: string;
  /** @nullable */
  transliteration: string | null;
  /** @nullable */
  translation?: string | null;
  difficulty: TongueTwisterDifficulty;
  isActive?: boolean;
  createdAt: Date;
}