# lib/api-zod/src/generated/types/tongueTwisterInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Punjabi Kids Gamification API
 * OpenAPI spec version: 0.1.0
 */
import type { TongueTwisterInputDifficulty } from './tongueTwisterInputDifficulty';

export interface TongueTwisterInput {
  text: string;
  transliteration?: string;
  translation?: string;
  difficulty: TongueTwisterInputDifficulty;
  isActive?: boolean;
}