# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Punjabi Kids Gamification API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  clerkId: string;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserRole;
  totalPoints: number;
  createdAt: Date;
}